Projects/AWS-Projects/AWS-3Tier-App/app/requirements.txt
flask
psycopg2-binary

#pip3 freeze > requirements.txt

